+++ js/data.js
@@ -1,0 +1,7 @@
+const LIST_MESSAGES = ['Всё отлично!', 'В целом всё неплохо.', 'Но не всё.', 'Когда вы делаете фотографию, хорошо бы убирать палец из кадра.', 'В конце концов это просто непрофессионально.', 'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.', 'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.', 'Лица у людей на фотке перекошены, как будто их избивают.', 'Как можно было поймать такой неудачный момент?!'];
+
+const SET_NAMES = ['Алексей', 'Наталья', 'Максим', 'Екатерина', 'Дмитрий', 'Ольга', 'Иван', 'Анна', 'Сергей', 'Юлия', 'Александра', 'Артем', 'Мария', 'Павел', 'Елена', 'Андрей', 'Татьяна', 'Владимир', 'Ирина', 'Константин', 'Анастасия', 'Роман', 'Светлана', 'Григорий', 'Виктория', 'Владислав', 'Людмила', 'Евгений', 'Олег', 'Алина'];
+
+const DESCRIPTION_PHOTOS = ['Пейзаж', 'Портрет', 'Горизонт', 'Закат', 'Цветы', 'Горы', 'Море', 'Город', 'Лес', 'Закрытые глаза', 'Улыбка', 'Семья', 'Еда', 'Дружба', 'Дети', 'Животные', 'Путешествие', 'Исследование', 'Спорт', 'Мода'];
+
+export{LIST_MESSAGES, SET_NAMES, DESCRIPTION_PHOTOS};
